Storybook@next + Webpack 5 + Yarn
This commit updates Storybook to 6.3+ and Webpack to 5+, for
local development specifically. We are not specifically using
Create React App yet, but when we do, we'll want to be conscious
that it prefers Webpack 4 for now. This should introduce no real
drift, as they're compile-time distinctions, but it means that
we may want to consider Webpack 5 based build pipelines over
CRA for when we progress to the deployment / distribution of
things.

Additionally, this commit throws out package-lock.json and prefers
yarn.lock instead; the bulk of everything resolves to the same
thing in the end, except version resolution / installs / script
runners get a bit easier to manage.