Absolute paths & package discipline changes. (#9)
This commit introduces a few new concepts for our consideration.

**Packages**

The first of these is the idea of "package discipline". Which is to say,
node / Javascript won't get in your way if you reach into packages and
grab whatever you want. This makes things tough to maintain and refactor,
down the line. It's _almost_ always better to ask the package to hand
its contents to you instead, meaning instead of this:

```tsx
import Button from '../../Button/Button'
```

This commit introduces a preference for this:

```tsx
import { Button } from 'components/Button'
```

We do this by updating the `tsconfig.json`, and our Storybook webpack
config, to consider `src` a module root for import resolution. This
means we've asked them to pretend "components" and the like are already
external modules.

The rules of package discipline are:

1. Relative imports are fine, inside a package. So, a file within
   `src/components/Button/` can grab a neighbor file:
  `./Button.styles.tsx`
2. Relative imports that go _up_ a level are not fine. So, going up
   into `src/components`, for example, is not:
  `'../Button/Button.styles.tsx'` is best avoided.
3. When importing, prefer a direct route to the package: for example,
   grabbing `components/Button` before `components`. This can help with
   bundle size and prevents circular imports.
4. If you wind up importing a _lot_ from a package, either consider
   organizing those imports so that the package contains all of them,
  _OR_ go ahead and pull from an index one directory up.

This leads to change 2:

**Index files**

Previously, we weren't using many index files, and so we had to reach
into neighbor components almost every time. Using index files, we can
give consumers a happy path for using the package's interface. This can
introduce a tension where we'll need to consider our interfaces as we
build and iterate, and generally that's a helpful tension, in my
experience.

**ESLint Rules*

The import groups have eslint rules around how they're structured, now.
I've tried to arrange it so that there will be auto-fixers, so if you
have a format on save routine set up, that should help. If not, run this
from your console to check:

```sh
yarn eslint src --fix
```

And that should apply any autofixes.

**Summary**

Most of these changes are made in the interest of aiding refactors as we
go. With explicit names, and a barrier that splits off knowledge of
where a file is located, I've found projects get much simpler to manage
and change over time. This specific commit should help declutter some of
our component space, for example, we can now refactor everything in the
`Table` components package without the rest of the app knowing. We also
hide our default exports behind deliberately named interfaces, meaning
project wide edits with refactor tools should be a little easier as
well.