Add registration workflow model to app state. (#17)
* Add registration workflow model to app state.

This commit introduces the registration workflow model to the app chart,
and exposes a few primitives through which to interact with it.

**OnState**

Any part of the application may render components by wrapping them in an
instance of the `<OnState />` component. It permits a typesafe way to
access the application state, much like a `<Route />` component in React
Router would render components based on location state.

For example, to render a spinner  when the app is checking for identity:

```ts
<OnState match="app.checking">
  <Spinner />
</OnState>
```

Along with the update to the app chart, this component will help
construct a workflow that ought to help prototyping somewhat.

**useAppState**

If you want to use the app state matcher directly, this helper hook can
assist you. This takes state matches down to a more terse, erognomic
form (hopefully).

**Updates to the State model**

The overall application itself is now more fully backed by a workflow
modeled out with xstate. Currently this looks for two properties, which
the `<Shell />` now accepts:

- `onAuthenticate`: A promise which resolves if a valid session already
  exists, or throws on a 401 or other errors.
- `onIdentity`: A promise which resolves if there is any identity info
  available for the application to use (for example if a provider has
  done the work to help set up a prospective caregiver for the app).

Right now these promises are no-op empty effects, but they can be
populated with dummy data for now, just as they can be eventually
updated to leverage actual network IO.

* Add xstate inspector.

* @wip