Integrate registration model and views. (#20)
This commit integrates the app chart and the registration views, but
keeps everything on `/registration` for now.

As an intermediary step, there is a page-specific hook designed to
automatically progress a viewer on the `/registration` page to the setup
wizard. This isn't meant for long term use at all - please, future us,
remember to remove this hook down the road. To help I've left plenty of
comments to remind us.