Move demo pages & add auth chart. (#48)
* Move existing pages to /demo.

In order to get moving on the authentication work for our registration
charter, we're beginning the deprecation of static pages. These will
each be reintroduced at some point when appropriate, but will be done so
test-first and in a dynamic fashion.

* Introduce auth machine.

This commit removes the preponderance of the registration chart,
relocating the wizard steps to a different node, and creating a parallel
auth node to handle our authentication flow.

Going forward, new portions of our state will adopt the parallel chart
approach. For example, forms may exist in the chart as distinct nodes in
the overall chart, or as sub-nodes of a "forms" state. Meanwhile, we'll
track overall form state in context.

For now, though, this sets us up for our first E2E tests and some
introductory auth systems.