Add identity form validation. (#52)
This commit introduces validation to the identity form part of our state
library, by integrating `yup`, and formalizing a model package where the
validations can be housed. Right now these validations are set up in a
way which the shell internally manages, but if we need to extract them,
later, we can simply provide validation through the current `services`
interface that `AppProvider` gives us.