Login page uses login form model. (#60)
This commit introduces a series of login form components, and leverages
them within the login page component. Right now, we're keeping these
components on the `/login` page, and we may use them with an external
consumer (auth 0), so we're holding off on full workflow integration for
now.

In addition, there's no organic way to surface the login form with the
current workflow (it relies heavily on variance in callbacks that we're
using placeholders for right now). So, for the time being, the login
page manually triggers some events. There's no reason to keep these once
the form is actually in a functional workflow.