Extensive testing/commenting of form factories. (#70)
First thing's first: this commit may look large, but I assure you, there
is very little new runtime code. Most of the changes are comments and
tests.

This commit changes no major functionality, but does significantly
change how we're adding form functionality. The notion is that adding a
form and its validation should not require work across the application
in terms of boilerplate, and that has largely been accomplished.
However, this commit pushes it further, cleaning up one remaining
boilerplate detail (app events in the AppProvider) and adding a
significant amount of test automation and JSDoc comments in the
meantime.

**To add a new form and validation.**

1. Create a new validation model; copying an existing one and modifying
   it is the best way to start. (Does plop make sense now?)
2. Make sure that model has been added to the `all` export in the index
   file for `app/state/forms/validation-models`.

And that should actually do it! The model, and all app events will be
provided for you in typesafe ways.

**As for the dev experience.**

- Hovering over a factory function should provide extensive summaries
  about what that function does.
- We have GENERATIVE tests that hinge on the `all` export, meaning each
  time a new model is added, that model is also added to major tests
  against the functionality of the factory suite.
- **Important**: One major issue with the current design is that it
  relies on named keywords and dynamic dispatch assembly to function.
  This is awesome because we can generate stuff, but it introduces the
  possibility of string collision. **There are now tests that will
  continually verify this collision has not happened.**