Broaden acceptance of StaticImageData types. (#96)
We've had a type gripe from CI which has (last minute) informed us that
Docker compilation won't accept `StaticImageData` for string fields. The
problem is, of course, it absolutely is fine.

This commit broadens our internal acceptance of StaticImageData in spots
where we'd formerly only accepted strings. This is sort of a hotfix, and
leaves us in a spot where we may still need to get these Docker-specific
errors earlier in our CI process.