Add patient GUID check to shell. (#100)
This commit adds a function to the shell that checks the url for the
presence of a Guid and Source query param. On the event of failure, the
shell will proceed along the current auth process. Since comprehension
of failure is baked in to the whole system so far, we should be able to
safely leverage this even in environments where the window might not be
present, so we go ahead and bake the functionality directly into the
default behavior of the shell.

The properties we care about, again, are: `Guid` and `Source`,
case-sensitive. So, this works: `path.to.app/?Guid=1234` but this
doesn't: `path.to.app/?guid=1234`.