Client consolidation. (#127)
This commit merges all of our async code into a consistent namespace,
surrounds them in tests, reduces the footprint of those async calls
in the rest of our application, and finally pushes a large amount of
our env config into a single package (`src/config`).

- There are two Client packages, Api and Services. Api clients use
  Next.js (local / public) apis, and use `fetch`. Service clients
  are meant to be used in back-end routes (privately), and leverage
  axios to connect different services.
- Where possible, we're going as terse as possible. For example,
  you should be able to insert `Client.Api.Identity.validate` directly
  into the Shell with no wrapping function. Similarly, the back-end
  functions should handle everything except building the response
  for the api handler.
- We've removed all default API responses. No mocks exist outside of
  tests.
- Most mocks and types, and type guards, have been moved to where
  they're most necessary. We import from `app` as little as possible.
  `src/client` should be parallel to `src/app`, not dependent on it.

This means from here on out, we'll want to be conscious about managing
our internal and external dependencies. If an external dependency
changes or exhibits new behavior, we will characterize it in a Client
test, then update our code. Similarly, if we need to update our own
api, we'll do that test first as well.

Other things:

- This commit also introduces the Validation API, for our registration
  form to check caregiver info against.
- There was a bug with the bleeding edge xstate inspector, so we've
  reverted to the default inspector for now.
- The auth flow was a little out of order: we were checking the guid
  before the session. Those have been flipped.