Metadata Profile integration with shell. (#146)
This commit introduces metadata integration with the shell. We'll be
combining this metadata with the information we get from the tests API
in order to construct the health profile. So, this is part 1 of a few
PRs which will accomplish that.

Adding this commit adds a dependency on some IDENTITY_* environment
variables, and so we also add those to Github and our deployment script.