Thread a caregiver's patientGuid through the app. (#152)
In order to get an improvement to the cohesion of registration / usage,
we're connecting the patient GUID directly to the Auth0 profile for a
caregiver. This means we'll have access to it wherever we have access to
a user session/profile, can pass it around less, and don't need to worry
as much about whether or not we're tracking it past registration.

A few things happen in here:

- We pass the patient GUID to the auth0 page, using a feature of the
  login handler.
- On the auth0 page, we now know whether or not we need to login or
  register based on the GUID presence. (No more register=true tricks.)
- We persist the patient GUID on the auth0 page as well.

I also noticed some page bloat happening in our Auth0 deployment, and
I've made a few changes to import statements to be more aggressive about
tree shaking.

Finally, we wire up a few caregiver values that'd been left unconnected.