Create app version of page layout. (#212)
The page layout had grown into something which was aware of some of our
app components. This led to a circular import issue where components was
requiring app, which required components. This commit splits off the
layout which needs awareness of the app, from the layout which doesn't,
but hopefully preserves the improved ergonomics we want of the page
layout component.