Reduce registration wizard renders. (#244)
This commit changes the registration wizard process in a way that
reduces the amount of distinct renders it invokes. Basically, instead of
rendering the same component a few times for different states, we now
render them once if any of the states match.

The idea here is simply to allow GTM tracking to avoid multiple calls.