Lessen registration syncing & last-step validation. (#274)
This commit changes our form sync feature (how we get prefills from
profile information) to be unidirectional instead of bidirectional.

Previously, whenever an identity form changed, it would change the
matching caregiver contact value, and vice versa. It turns out, though,
that making a call to `/api/identity/confirm` (backed by the Provider
portal's `/UpdateRegistrationStatus`) expects the originating values.

Or, in other words, if you edit contact info later, you can't register.

Making it unidirectional, and then calling `confirm` with the identity
payload works.

Because this is all sorted in this commit, we also remove the second
validation call we executed before attempting to register, as it was
only there to get us better error information.